Initial move of SW functions.
